moved json save/load to the project method
