+++ RATapi/utils/convert.py
@@ -1,6 +1,5 @@
 """Utilities for converting input files to Python `Project`s."""
 
-import json
 import warnings
 from collections.abc import Iterable
 from os import PathLike
@@ -553,72 +552,3 @@ def project_class_to_r1(
     eng.save(str(filename), "problem", nargout=0)
     eng.exit()
     return None
-
-
-def project_to_json(project: Project) -> str:
-    """Write a Project as a JSON file.
-
-    Parameters
-    ----------
-    project : Project
-        The input Project object to convert.
-
-    Returns
-    -------
-    str
-        A string representing the class in JSON format.
-    """
-    json_dict = {}
-    for field in project.model_fields:
-        attr = getattr(project, field)
-
-        if field == "data":
-
-            def make_data_dict(item):
-                return {
-                    "name": item.name,
-                    "data": item.data.tolist(),
-                    "data_range": item.data_range,
-                    "simulation_range": item.simulation_range,
-                }
-
-            json_dict["data"] = [make_data_dict(data) for data in attr]
-
-        elif field == "custom_files":
-
-            def make_custom_file_dict(item):
-                return {"name": item.name, "filename": item.filename, "language": item.language, "path": str(item.path)}
-
-            json_dict["custom_files"] = [make_custom_file_dict(file) for file in attr]
-
-        elif isinstance(attr, ClassList):
-            json_dict[field] = [dict(item) for item in attr]
-        else:
-            json_dict[field] = attr
-
-    return json.dumps(json_dict)
-
-
-def project_from_json(input: str) -> Project:
-    """Read a Project from a JSON string generated by `to_json`.
-
-    Parameters
-    ----------
-    input : str
-        The JSON input as a string.
-
-    Returns
-    -------
-    Project
-        The project corresponding to that JSON input.
-    """
-    model_dict = json.loads(input)
-    for i in range(0, len(model_dict["data"])):
-        if model_dict["data"][i]["name"] == "Simulation":
-            model_dict["data"][i]["data"] = empty([0, 3])
-            del model_dict["data"][i]["data_range"]
-        else:
-            data = model_dict["data"][i]["data"]
-            model_dict["data"][i]["data"] = array(data)
-
-    return Project.model_validate(model_dict)
